refactored layouts
Moved options which define the chart and not the dataset to chart options.
